abstact singletons added, carMovementController & playerInput dependency fixed
