+++ Fast-and-fractured/Assets/FastAndFractured/!Prototype/Scripts/LevelController.cs
@@ -3,21 +3,60 @@ using System.Collections.Generic;
 using StateMachine;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using Utilities;
 
 namespace Game { 
-public class LevelController : MonoBehaviour
+public class LevelController : AbstractSingleton<LevelController>
 {
-       
+    public bool UsingController;
     [SerializeField] private List<ObjectPoolSO> poolSOList;
     [SerializeField] private List<StatsController> characters;
     [SerializeField] private EnemyAIBrain ai;
     [SerializeField] private List<KillCharacterNotify> killCharacterHandles;
     [SerializeField] private List<Controller> controllers;
     // Start is called before the first frame update
+    protected override void Awake()
+    {
+        base.Awake();   
+    }
+
     void Start()
     {
         StartLevel();
     }
+    // this will be moved to gameManaager once its created
+
+    private void OnEnable()
+    {
+        PlayerInputController.OnInputDeviceChanged += HandleInputChange;
+    }
+
+    private void OnDisable()
+    {
+        PlayerInputController.OnInputDeviceChanged -= HandleInputChange;
+    }
+
+    public void HandleInputChange(INPUT_DEVICE_TYPE inputType)
+    {  
+        if (inputType == INPUT_DEVICE_TYPE.KeyboardMouse)
+        {
+            UsingController = false;
+        }
+        else if (inputType == INPUT_DEVICE_TYPE.XboxController || inputType == INPUT_DEVICE_TYPE.PSController)
+        {
+            UsingController = true;
+        }
+
+        foreach (StatsController character in characters) // for now when inoput changed its detected it will only notify the carMovementController of the player
+        {
+            if (character.CompareTag("Player"))
+            {
+                    character.gameObject.GetComponent<CarMovementController>().HandleInputChange(UsingController);
+            }
+        }
+
+    }
+    // will be moved to gameManager
 
     // Update is called once per frame
     void Update()
