+++ Fast-and-fractured/Assets/Utilities/AbstractSingleton.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Utilities
